// cypress/support/enums.ts
export enum PAGES {
    INVESTOR_OVERVIEW = './investor-overview.html',
    MATCHING = './Matching.html',
}

export enum LOCATOR {
    ADD_INVESTOR_BTN = 'add-investor',
    CLOSING = 'closing',
    ENTITY = 'entity',
    FILTER_BTN = 'filter-btn',
    SAVE_INVESTOR_BTN = 'save-investor',
    START_MATCHING_BTN = 'start-matching',
    STATUS = 'status',
    SUGGESTED = 'suggested',
    VISIBILITY_CHECKBOX = 'visibility-checkbox'
}

export enum ENTITIES {
    NEW_ENTITY = "Epsilon Funds"
}

export enum CHECKBOXES {
    CHECKED = "✅",
    NOT_CHECKED = "⬜",
}